requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
blis==0.7.9
Bottleneck==1.3.5
cachetools==5.2.0
catalogue==2.0.8
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
confection==0.0.3
contourpy==1.0.6
cycler==0.11.0
cymem==2.0.7
deepface==0.0.75
dnspython==2.3.0
dpcpp-cpp-rt==2023.0.0
email-validator==1.3.0
fastai==1.0.57
fastapi==0.89.1
fastprogress==1.0.3
filelock==3.9.0
fire==0.5.0
Flask==2.2.2
flatbuffers==22.12.6
fonttools==4.38.0
frozenlist==1.3.3
fst-pso==1.8.1
FuzzyTM==2.0.5
gast==0.4.0
gdown==4.6.0
gensim==4.3.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
gunicorn==20.1.0
h11==0.14.0
h5py==3.7.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
imageio==2.24.0
importlib-metadata==6.0.0
inltk==0.9
intel-cmplr-lib-rt==2023.0.0
intel-cmplr-lic-rt==2023.0.0
intel-opencl-rt==2023.0.0
intel-openmp==2023.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
langcodes==3.3.0
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
miniful==0.0.6
mkl==2023.0.0
mkl-fft==1.3.1
mkl-service==2.4.0
mtcnn==0.1.1
multidict==6.0.4
murmurhash==1.0.9
networkx==3.0
numexpr==2.8.4
numpy==1.22.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-ml-py3==7.352.0
oauthlib==3.2.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
orjson==3.8.5
packaging==23.0
pandas==1.5.2
pathy==0.10.1
Pillow==9.4.0
preshed==3.0.8
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.4
pyFUME==0.2.25
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==19.0.2
redis==4.4.2
requests==2.28.1
requests-oauthlib==1.3.1
retina-face==0.0.12
rfc3986==1.5.0
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.2
sentencepiece==0.1.97
simpful==2.9.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.4.4
spacy-legacy==3.0.11
spacy-loggers==1.0.4
spotipy==2.22.0
srsly==2.4.5
starlette==0.22.0
tbb==2021.8.0
tekore==4.5.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
thinc==8.1.6
threadpoolctl==3.1.0
tifffile==2022.10.10
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
typer==0.7.0
typing==3.7.4.3
typing_extensions==4.4.0
ujson==5.7.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
wasabi==0.10.1
watchfiles==0.18.1
websockets==10.4
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.2
zipp==3.11.0
